Create README.md
In this study, a unified spectral-spatial channel grouping based swin transformer adaptation is proposed. Our model explores the connection between the spectral correlation and spatial correlation information from the grouping of channels based on the spectral reflectivity. The product of channel grouping using a shifted window based transformer, results in a richer embedding. This can be exploited to achieve better visual quality performance.